+++ src/main/java/com/jj/Gradebook/dao/AnnouncementRepository.java
@@ -3,5 +3,5 @@ package com.jj.Gradebook.dao;
 import com.jj.Gradebook.entity.Announcements;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnnouncementRepository extends JpaRepository<Announcements,Integer> {
+public interface AnnouncementRepository extends JpaRepository<Announcements,Long> {
 }
