+++ src/main/java/com/jj/Gradebook/dao/GradeRepository.java
@@ -3,5 +3,5 @@ package com.jj.Gradebook.dao;
 import com.jj.Gradebook.entity.Grade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GradeRepository extends JpaRepository<Grade, Integer> {
+public interface GradeRepository extends JpaRepository<Grade, Long> {
 }
