+++ src/main/java/com/jj/Gradebook/dao/StudentRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StudentRepository extends JpaRepository<Student,Integer> {
+public interface StudentRepository extends JpaRepository<Student,Long> {
     Optional<Student> findStudentByUserPesel(String pesel);
+
 }
