+++ src/main/java/com/jj/Gradebook/service/message/MessageService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface MessageService {
     List<Message> findAll();
-    Message findById(int id);
+    Message findById(Long id);
     Message save(Message message);
-    void deleteById(int id);
+    void deleteById(Long id);
 }
