+++ src/main/java/com/jj/Gradebook/service/note/NoteService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface NoteService {
     List<Note> findAll();
-    Note findById(int id);
+    Note findById(Long id);
     Note save(Note note);
-    void deleteById(int id);
+    void deleteById(Long id);
 }
